Add initial version of top level macros
